projects/Face_Recognition-main/requirements.txt
opencv-contrib-python
numpy
Pillow
streamlit
pandas
